app/requirements.txt
confluent-kafka
python-dotenv
tqdm
